fix : int change to long

+++ src/main/java/com/chaffee/mapper/UserMapper.java
@@ -22,7 +22,7 @@ public interface UserMapper extends BaseMapper<User> {
   LoginDTO queryLogin( @Param( "userCode" ) String userCode );
   
   List<UserVO> queryUserList( @Param( "page" ) IPage<UserVO> page, @Param( "userName" ) String userName, @Param(
-      "userRole" ) int userRole );
+      "userRole" ) long userRole );
   
   UserVO queryUser(@Param( "id" )long id);
